+++ tests/postgres_test.go
@@ -136,3 +136,14 @@ func TestPostgresHook(t *testing.T) {
 		t.Fatalf("query failed: %s", err)
 	}
 }
+
+// Test to verify that when a proxy client serves one postgres instance that can be
+// dialed successfully, the health check readiness endpoint serves http.StatusOK.
+func TestPostgresDial(t *testing.T) {
+	switch "" {
+	case *postgresConnName:
+		t.Fatal("'postgres_conn_name' not set")
+	}
+
+	singleInstanceDial(t, *postgresConnName, postgresPort)
+}